refactor api call
