+++ ecommerce/src/components/Dashboard/Card.js
@@ -1,32 +1,19 @@
 import "./Card.css";
 import React, { useEffect, useState, useContext } from "react";
-import { db } from "../../config/Firebase";
-import { getDocs, collection } from "firebase/firestore";
 import { CartContext } from "../../App";
-
+import { getProductList} from "../../FirebaseCall";
 const Card = () => {
   const [productsList, setProductsList] = useState([]);
   const cartContext = useContext(CartContext);
 
-  const productsCollectionRef = collection(db, "products");
-
-  useEffect(() => {
-    const getProductList = async () => {
-      try {
-        const data = await getDocs(productsCollectionRef);
-        const filteredData = data.docs.map((doc) => ({
-          ...doc.data(),
-          id: doc.id,
-        }));
-
-        console.log(filteredData);
-        setProductsList(filteredData);
-      } catch (err) {
-        console.error(err);
-      }
-    };
-    getProductList();
-  }, []);
+  useEffect(()=>{
+    getProductList().then((filteredData)=>{
+      setProductsList(filteredData)
+    })
+  
+  
+    
+  },[])
 
   const handleAddToCart = (product) => {
     cartContext.addToCart(product);
@@ -34,16 +21,20 @@ const Card = () => {
 
   return (
     <>
-      {productsList.map((product, index) => (
-        <div className="card-div" key={index}>
-          <div>
-            <p>{product.MARCA}</p>
-            <p>{product.PRODUCTO}</p>
-            <p>${product.PRECIO}</p>
-          </div>
-          <button onClick={() => handleAddToCart(product)}>
-            Agregar al carrito
-          </button>
+      {productsList.map((product) => (
+        <div key={product.id}>
+          {product.STATE && (
+            <div className="card-div" >
+              <div>
+                <p>{product.MARCA}</p>
+                <p>{product.PRODUCTO}</p>
+                <p>${product.PRECIO}</p>
+              </div>
+              <button onClick={() => handleAddToCart(product)}>
+                Agregar al carrito
+              </button>
+            </div>
+          )}
         </div>
       ))}
     </>
